refactored method generating group links
