+++ public/src/components/teacher/GroupResults.tsx
@@ -2,8 +2,7 @@ import * as React from "react";
 import { Assignment, Course } from "../../../proto/ag_pb";
 import { DynamicTable, Row, Search, StudentLab } from "../../components";
 import { IAssignmentLink, IStudentSubmission } from "../../models";
-import { sortByScore } from "./sorter";
-import { slugify } from "../../helper"
+import { generateGroupRepoLink, sortByScore } from "./groupHelper";
 
 interface IResultsProps {
     course: Course;
@@ -95,7 +94,7 @@ export class GroupResults extends React.Component<IResultsProps, IResultsState> 
     private getGroupResultSelector(group: IAssignmentLink): Array<string | JSX.Element> {
         const slipdayPlaceholder = "5";
         const grp = group.link.getGroup();
-        const name = grp ? this.generateGroupRepoLink(grp.getName()) : "";
+        const name = grp ? generateGroupRepoLink(grp.getName(), this.props.courseURL) : "";
         let selector: Array<string | JSX.Element> = [name, slipdayPlaceholder];
         selector = selector.concat(group.assignments.filter((e) => e.assignment.getIsgrouplab()).map((e) => {
             let approvedCss;
@@ -121,10 +120,6 @@ export class GroupResults extends React.Component<IResultsProps, IResultsState> 
                 authorName: item.authorName,
             });
         }
-    }
-
-    private generateGroupRepoLink(groupname: string): JSX.Element {
-        return <a href={this.props.courseURL + slugify(groupname)}>{ groupname }</a>;
     }
 
     private handleOnchange(query: string): void {
